move to f32 for rasters

+++ src/consts.rs
@@ -5,7 +5,7 @@ pub const INV_CELL_SIZE_METERS_USIZE: usize = 2; // test 1, 2 or 4
 pub const SIMPLIFICATION_DIST: f64 = 0.1;
 pub const MERGE_DELTA: f64 = 0.1;
 
-pub const MIN_GRAD_LENGTH: f64 = 1.0;
+pub const MIN_GRAD_LENGTH: f32 = 1.0;
 
 pub const TILE_SIZE_PIXELS: usize = INV_CELL_SIZE_METERS_USIZE * TILE_SIZE_METERS_USIZE;
 pub const CELL_SIZE_METERS: f64 = 1. / INV_CELL_SIZE_METERS_USIZE as f64;
